+++ Demos/CFlexray/FlexrayDemo.c
@@ -502,7 +502,6 @@ int main(int argc, char** argv)
         printf("SilKit_Participant_Create => %s\n", SilKit_GetLastErrorString());
         return 2;
     }
-    printf("Creating participant '%s' for simulation '%s'\n", participantName, registryUri);
 
     SilKit_LifecycleService* lifecycleService;
     returnCode = SilKit_LifecycleServiceWithTimeSync_Create(&lifecycleService, participant);