+++ SilKit/include/silkit/capi/SilKit.h
@@ -62,7 +62,7 @@ SilKitAPI const char* SilKit_GetLastErrorString();
 
 typedef  const char*(*SilKit_GetLastErrorString_t)();
 
-typedef SilKit_ReturnCode(*SilKit_Participant_CreateLogger_t)(
+typedef SilKit_ReturnCode(*SilKit_Participant_GetLogger_t)(
     SilKit_Logger** outLogger,
     SilKit_Participant* participant);
 
@@ -75,7 +75,7 @@ typedef SilKit_ReturnCode(*SilKit_Participant_CreateLogger_t)(
  * There is no futher cleanup necessary, except for destroying the simulation participant at the end of the
  * simulation.
  */
-SilKitAPI SilKit_ReturnCode SilKit_Participant_CreateLogger(
+SilKitAPI SilKit_ReturnCode SilKit_Participant_GetLogger(
     SilKit_Logger** outLogger,
     SilKit_Participant* participant);