+++ Utilities/SilKitSystemMonitor/PassiveSystemMonitor.cpp
@@ -123,7 +123,7 @@ int main(int argc, char** argv)
 
         auto participant = SilKit::CreateParticipant(std::move(configuration), participantName, connectUri);
 
-        auto* logger = participant->CreateLogger();
+        auto* logger = participant->GetLogger();
         auto* systemMonitor = participant->CreateSystemMonitor();
 
         systemMonitor->AddParticipantStatusHandler([logger](const Services::Orchestration::ParticipantStatus& status) {